requirements.txt
torch==2.4.0
torchvision==0.19.0
matplotlib==3.9.2
Flask==3.0.3
gunicorn
Flask==3.0.3
PyYAML==6.0.2
scikit-learn==1.5.1

